+++ src/Alpine.jl
@@ -13,9 +13,10 @@ const Alp = Alpine
 include("const.jl")
 
 # Engine for high-level algorithmic control and user-interface
-include("matrix_opt_interface.jl")
-include("moi_function2expr.jl")
+
 include("solver.jl")
+include("MOI_wrapper/MOI_wrapper.jl")
+include("MOI_wrapper/MOI_function2expr.jl")
 
 # Engine for expression handling
 include("nlexpr.jl")